# docs/design/openapi.yaml
openapi: 3.1.1
...
info:
  title: Infrastructure Platform Execution API
  version: 0.2.0
  description: >-
    M0 target contract, not a claim of local implementation conformance. The
    local core increment implements a subset with selectable fixture identities;
    see ../progress.md for gaps. M1 exercises the fake provider; M3 supplies the
    selected live provider. This target requires API-audience Entra access tokens.
    Object/data authorization is enforced on every
    read and cancellation. All operations accept validated X-Flow-ID and W3C
    trace context and return server X-Request-ID and X-Flow-ID headers; neither
    grants authority. Local conventions await enterprise review V01.
  contact:
    name: Platform engineering owners (assignment pending)
  x-api-id: infrastructure-platform-execution
  x-audience: company-internal
...
servers:
  - url: https://platform.example.com
    description: Documentation placeholder; not a deployed service.
security:
  - EntraBearer: []
...
paths:
  /identity-context:
    get:
      operationId: getIdentityContext
      tags: [Identity]
      summary: Read the caller's authorized application/environment context
      parameters:
        - { $ref: '#/components/parameters/FlowId' }
        - { $ref: '#/components/parameters/Traceparent' }
        - { $ref: '#/components/parameters/Tracestate' }
      responses:
        '200':
          description: Caller context only; grants are evaluated again on each operation.
          headers:
            X-Flow-ID: { $ref: '#/components/headers/FlowId' }
            Cache-Control: { $ref: '#/components/headers/NoStore' }
            X-Request-ID: { $ref: '#/components/headers/RequestId' }
          content:
            application/json:
              example: {"principal_id":"principal-example01","principal_kind":"human","authorized_contexts":[{"application_id":"software-factory","environment":"development","permissions":["execution.submit","execution.read","execution.cancel","execution.data.read","catalog.read"]}]}
              schema: { $ref: '#/components/schemas/IdentityContext' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '400': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '406': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/RetryableProblem' }
        '503': { $ref: '#/components/responses/RetryableProblem' }
        default: { $ref: '#/components/responses/Problem' }
  /execution-templates:
    get:
      operationId: listExecutionTemplates
      security:
        - EntraBearer: [catalog.read]
      x-required-permissions: [catalog.read]
      tags: [Templates]
      summary: List authorized immutable execution-template versions
      description: Authorization filtering precedes pagination; no global count is returned.
      parameters:
        - { $ref: '#/components/parameters/FlowId' }
        - { $ref: '#/components/parameters/Traceparent' }
        - { $ref: '#/components/parameters/Tracestate' }
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: Authorized template page.
          headers:
            X-Request-ID: { $ref: '#/components/headers/RequestId' }
            X-Flow-ID: { $ref: '#/components/headers/FlowId' }
            Cache-Control: { $ref: '#/components/headers/NoStore' }
          content:
            application/json:
              example: {"items":[{"template_id":"pr-validation-v1","version":"1.0.0","description":"Source tests against a supervisor-started pinned helper service; emits named sbom. Command is derived from the single input reference.","allowed_overrides":["compute_profile","timeout_seconds"],"input_schema":{"$schema":"https://json-schema.org/draft/2020-12/schema","type":"object","required":["application_id","environment","template_id","template_version","input_artifact_refs"],"properties":{"template_id":{"const":"pr-validation-v1"},"template_version":{"const":"1.0.0"},"compute_profile":{"enum":["general-medium"]},"timeout_seconds":{"type":"integer","minimum":1,"maximum":1800},"input_artifact_refs":{"type":"array","minItems":1,"maxItems":1,"items":{"type":"string"}},"image":{"const":"registry.example.com/software-factory/reviewer@sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"},"execution_class":{"const":"isolated-vm"},"placement_policy":{"const":"us-data-residency"},"network_profile":{"const":"software-factory-restricted"},"environment_variables":{"const":{"RUN_MODE":"pull-request"}},"secret_refs":{"const":[]}},"description":"Template-specific constraints apply with the closed structural ExecutionInput schema and server-derived command binding; these annotations do not execute expansion."},"defaults":{"image":"registry.example.com/software-factory/reviewer@sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa","compute_profile":"general-medium","execution_class":"isolated-vm","placement_policy":"us-data-residency","network_profile":"software-factory-restricted","timeout_seconds":1800,"environment_variables":{"RUN_MODE":"pull-request"},"secret_refs":[]},"revoked":false}],"page":{}}
              schema:
                type: object
                required: [items, page]
                properties:
                  items:
                    type: array
                    maxItems: 200
                    items: { $ref: '#/components/schemas/ExecutionTemplate' }
                  page: { $ref: '#/components/schemas/Page' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '400': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '406': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/RetryableProblem' }
        '503': { $ref: '#/components/responses/RetryableProblem' }
        default: { $ref: '#/components/responses/Problem' }
  /executions:
    post:
      operationId: createExecution
      security:
        - EntraBearer: [execution.submit]
      tags: [Executions]
      summary: Durably admit one template-bound execution
      description: >-
        Requires execution.submit on the registered application/environment.
        All fields are governed by the selected immutable template. Ownership,
        targets, profile versions and identities are resolved by the server.
        Identical authorized retries return the original 202 and Location.
        Payload conflicts return 409. Acceptance does not mean compute started.
      x-required-permissions: [execution.submit]
      parameters:
        - { $ref: '#/components/parameters/Traceparent' }
        - { $ref: '#/components/parameters/Tracestate' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - { $ref: '#/components/parameters/FlowId' }
      requestBody: { $ref: '#/components/requestBodies/ExecutionInput' }
      responses:
        '202': { $ref: '#/components/responses/ExecutionAccepted' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '413': { $ref: '#/components/responses/Problem' }
        '415': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/RetryableProblem' }
        '503': { $ref: '#/components/responses/RetryableProblem' }
        '404': { $ref: '#/components/responses/Problem' }
        '406': { $ref: '#/components/responses/Problem' }
        '410': { $ref: '#/components/responses/Problem' }
        default: { $ref: '#/components/responses/Problem' }
  /execution-templates/{template_id}/executions:
    parameters:
      - name: template_id
        in: path
        required: true
        schema: { $ref: '#/components/schemas/LogicalId' }
    post:
      operationId: createTemplateExecution
      security:
        - EntraBearer: [execution.submit]
      tags: [Templates, Executions]
      summary: Submit through a template-scoped route
      description: >-
        Identical to POST /executions. Body template_id must match the path.
        Both routes normalize to the same submission operation for idempotency.
      x-required-permissions: [execution.submit]
      parameters:
        - { $ref: '#/components/parameters/Traceparent' }
        - { $ref: '#/components/parameters/Tracestate' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - { $ref: '#/components/parameters/FlowId' }
      requestBody: { $ref: '#/components/requestBodies/ExecutionInput' }
      responses:
        '202': { $ref: '#/components/responses/ExecutionAccepted' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/RetryableProblem' }
        '503': { $ref: '#/components/responses/RetryableProblem' }
        '406': { $ref: '#/components/responses/Problem' }
        '410': { $ref: '#/components/responses/Problem' }
        '413': { $ref: '#/components/responses/Problem' }
        '415': { $ref: '#/components/responses/Problem' }
        default: { $ref: '#/components/responses/Problem' }
  /executions/{execution_id}:
    parameters:
      - { $ref: '#/components/parameters/ExecutionId' }
    get:
      operationId: getExecution
      security:
        - EntraBearer: [execution.read]
      tags: [Executions]
      summary: Read execution status and independent cleanup/delivery outcomes
      x-required-permissions: [execution.read]
      parameters:
        - { $ref: '#/components/parameters/FlowId' }
        - { $ref: '#/components/parameters/Traceparent' }
        - { $ref: '#/components/parameters/Tracestate' }
        - name: If-None-Match
          in: header
          schema: { type: string, maxLength: 128 }
      responses:
        '200':
          description: Current authorized projection; provider identifiers are never exposed.
          headers:
            X-Flow-ID: { $ref: '#/components/headers/FlowId' }
            ETag: { $ref: '#/components/headers/ETag' }
            Cache-Control: { $ref: '#/components/headers/NoStore' }
            X-Request-ID: { $ref: '#/components/headers/RequestId' }
          content:
            application/json:
              example: {"id":"exec_example01","application_id":"software-factory","environment":"development","template_id":"pr-validation-v1","template_version":"1.0.0","state":"succeeded","cleanup_state":"succeeded","delivery_status":"complete","infrastructure_status":"absent","dispatch_status":"started","result_complete":true,"revision":12,"created_at":"2026-09-05T18:00:00Z","deadline_at":"2026-09-05T18:30:00Z","started_at":"2026-09-05T18:01:00Z","observed_at":"2026-09-05T18:02:00Z","completed_at":"2026-09-05T18:01:45Z","links":{"self":"https://platform.example.com/executions/exec_example01","events":"https://platform.example.com/executions/exec_example01/events","logs":"https://platform.example.com/executions/exec_example01/logs","results":"https://platform.example.com/executions/exec_example01/results"}}
              schema: { $ref: '#/components/schemas/Execution' }
        '304':
          description: Unchanged after current object authorization.
          headers:
            X-Request-ID: { $ref: '#/components/headers/RequestId' }
            X-Flow-ID: { $ref: '#/components/headers/FlowId' }
            ETag: { $ref: '#/components/headers/ETag' }
            Cache-Control: { $ref: '#/components/headers/NoStore' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/Problem' }
        '400': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '406': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/RetryableProblem' }
        '503': { $ref: '#/components/responses/RetryableProblem' }
        default: { $ref: '#/components/responses/Problem' }
  /executions/{execution_id}/cancellations:
    parameters:
      - { $ref: '#/components/parameters/ExecutionId' }
    post:
      operationId: requestExecutionCancellation
      security:
        - EntraBearer: [execution.cancel]
      tags: [Executions]
      summary: Record an eventual stop request
      description: >-
        Requires current cancellation authority. No rollback/removal guarantee.
        The execution may complete before the stop request takes effect.
        New intents on terminal executions return 409; idempotent retries replay
        the original response. A cancellation resource does not require If-Match.
      x-required-permissions: [execution.cancel]
      parameters:
        - { $ref: '#/components/parameters/Traceparent' }
        - { $ref: '#/components/parameters/Tracestate' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - { $ref: '#/components/parameters/FlowId' }
      requestBody:
        required: false
        description: An absent body normalizes to an empty object for idempotency.
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                reason:
                  type: string
                  enum: [no_longer_needed, superseded_by_new_request, operator_request]
            example: { reason: no_longer_needed }
      responses:
        '202':
          description: Durable intent; Location is the execution status resource.
          headers:
            X-Request-ID: { $ref: '#/components/headers/RequestId' }
            X-Flow-ID: { $ref: '#/components/headers/FlowId' }
            Location: { $ref: '#/components/headers/Location' }
            Retry-After: { $ref: '#/components/headers/RetryAfter' }
            Cache-Control: { $ref: '#/components/headers/NoStore' }
          content:
            application/json:
              example: {"id":"cancel-example01","execution_id":"exec_example01","status":"requested","requested_at":"2026-09-05T18:01:00Z","deadline_at":"2026-09-05T18:03:00Z","reason":"no_longer_needed"}
              schema: { $ref: '#/components/schemas/Cancellation' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '400': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '406': { $ref: '#/components/responses/Problem' }
        '410': { $ref: '#/components/responses/Problem' }
        '413': { $ref: '#/components/responses/Problem' }
        '415': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/RetryableProblem' }
        '503': { $ref: '#/components/responses/RetryableProblem' }
        default: { $ref: '#/components/responses/Problem' }
  /executions/{execution_id}/events:
    parameters:
      - { $ref: '#/components/parameters/ExecutionId' }
    get:
      operationId: getExecutionEvents
      security:
        - EntraBearer: [execution.read]
      tags: [Executions]
      summary: Poll ordered execution events with an optional bounded wait
      description: >-
        Reauthorize before returning after a wait. Empty timeout is 200 with no
        items and the same cursor. Events are deduplicated by execution/sequence.
        Expired cursor returns 410. Disconnect does not cancel the execution.
      x-required-permissions: [execution.read]
      parameters:
        - { $ref: '#/components/parameters/FlowId' }
        - { $ref: '#/components/parameters/Traceparent' }
        - { $ref: '#/components/parameters/Tracestate' }
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
        - name: wait_seconds
          in: query
          schema: { type: integer, format: int32, minimum: 0, maximum: 25, default: 0 }
      responses:
        '200':
          description: Authorized events in ascending sequence order.
          headers:
            X-Request-ID: { $ref: '#/components/headers/RequestId' }
            X-Flow-ID: { $ref: '#/components/headers/FlowId' }
            Cache-Control: { $ref: '#/components/headers/NoStore' }
          content:
            application/json:
              example: {"items":[{"id":"event-example12","execution_id":"exec_example01","sequence":12,"event_type":"execution.cleanup_changed","recorded_at":"2026-09-05T18:02:01Z","observed_at":"2026-09-05T18:02:00Z","revision":12,"state":"succeeded","cleanup_state":"succeeded","delivery_status":"complete","infrastructure_status":"absent","dispatch_status":"started","result_complete":true}],"cursor":"opaque-event-cursor-12","page":{}}
              schema:
                type: object
                required: [items, cursor, page]
                properties:
                  items:
                    type: array
                    maxItems: 200
                    items: { $ref: '#/components/schemas/ExecutionEvent' }
                  cursor: { type: string, maxLength: 2048 }
                  page: { $ref: '#/components/schemas/Page' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/Problem' }
        '410': { $ref: '#/components/responses/Problem' }
        '400': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '406': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/RetryableProblem' }
        '503': { $ref: '#/components/responses/RetryableProblem' }
        default: { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    EntraBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Entra access token for this API; verify configured tenant/issuer/audience,
        signature and time claims. Humans use delegated permissions obtained by
        authorization code with PKCE; CI uses federated app identity and app roles.
        ID tokens and management tokens are not accepted. OAuth endpoints and
        exact scopes/roles are tenant-registration decisions V02, not placeholders
        for a new token issuer. Operation permission extensions supplement this
        transport scheme; application/object grants are mandatory. Per-operation
        security arrays list required registry permission roles (OpenAPI 3.1),
        not substitutes for tenant-issued scopes or app roles.
  parameters:
...
    FlowId:
      name: X-Flow-ID
      in: header
      required: false
      description: Untrusted correlation only; validate or generate when absent. Not an execution ID or credential.
      schema: { type: string, pattern: '^[a-zA-Z0-9/+_=\-]{1,128}$' }
    Traceparent:
      name: traceparent
      in: header
      description: Validated W3C trace context; untrusted correlation only.
      schema: { type: string, maxLength: 512 }
    Tracestate:
      name: tracestate
      in: header
      description: Allowlisted W3C trace state; never authorization or arbitrary baggage.
      schema: { type: string, maxLength: 512 }
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor bound to resource, scope, ordering and expiry; always reauthorize.
      schema: { type: string, maxLength: 2048 }
...
  headers:
...
    RequestId:
      description: Server-generated request correlation ID.
      schema: { type: string, maxLength: 128 }
    FlowId:
      description: Validated or server-generated flow correlation ID.
      schema: { type: string, maxLength: 128 }
...
    NoStore:
      description: Authenticated responses are not shared-cacheable.
      schema: { type: string, const: no-store }
...
  schemas:
...
    IdentityContext:
      type: object
      required: [principal_id, principal_kind, authorized_contexts]
      properties:
        principal_id: { $ref: '#/components/schemas/LogicalId' }
        principal_kind: { type: string, examples: [human, application] }
        authorized_contexts:
          type: array
          maxItems: 200
          description: M1 enrollment bounded to 200 contexts per principal; reject broader onboarding until pagination is designed.
          items:
            type: object
            required: [application_id, environment, permissions]
            properties:
              application_id: { $ref: '#/components/schemas/LogicalId' }
              environment: { $ref: '#/components/schemas/LogicalId' }
              permissions:
                type: array
                maxItems: 32
                items: { type: string, maxLength: 64 }
...